cpufreq: schedutil: Remove all iowait boosting code
There is little benefit to increasing CPU frequencies when we are
waiting on our flash chip to finish up.

I have seen commits that make iowait boosting optional, but I would like
to erase all traces of the boosting as to avoid any unnecessary
timestamping or overhead in the governor.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Forenche <prahul2003@gmail.com>
Change-Id: Id2cb55bd0f37760968d619d50e86d58f3a7beba3